stable spot. Probably should start documenting tests. Have not implemented the number of passes option yet in the menu.c (or added it to the driver.c), this is easily a counter which just rewinds the file and calls compress again. Decompression could also have a number passes option. The number of passes for compression/decompression should obviously match, no reason to store this information in the compressed file (if it even really has benefit, I've seen the second/third pass may slightly reduce the size but not always, then it always starts going back up, maybe it could benefit encryption or obfuscation) I have alternate code in the Huffman tree builder that accomplishes reinserting nodes back into the min heap, the filesizes I saw before were larger using this so I commented it out. I tried storing the maximum bits needed to represent the largest byte to reduce the bits used to represent each byte. I think this is typically not helpful and adds a byte. Could reverse the zero / one encodings, to where zero is right and one is left. Reversing the encodings could be helpful in combination with the node reinsertion logic. If trees get taller, but the tall path is zeroes we can end up with more zeroes which could help recompression.

+++ huffmantree.c
@@ -674,7 +674,7 @@ void docompress(FILE *infile,FILE *outfile,const byte doprints) {
 	if (bitshmt<0) { bitshmt+=8; writebuffer[++bufferpos]=0; }
 	// for all unique bytes
 	while (bytesremaining>0) {
-		byte groupcount = 0;
+		int groupcount = 0; // worst case the group has 256 ...
 		//for each byte + encoding . . .
 		for (i=0;i<uniquebytes;++i) {
 			// group by lengths/level
